Menu/UX: Indicate that SAE Service Level 1 is meant for 120V connections only
To help steer non-American users away from selecting Level 1
and having their Wh calculations wrong (due to VOLTS_FOR_L1
that's used in KWH_RECORDING being hardcoded to 120V).
I.e. people who select Level 1 but are running off 240V mains
are presented with kWh display that's half of the actual usage.

+ move the alternative commented out define for 230V L2 voltage after
the default 240V L2 definition, so that the 230V is not mistakenly read as being an alternative for L1

http://www.sae.org/smartgrid/chargingspeeds.pdf