Fixing bug where -0.5c was reported as -256C
This is a fix to my "fix".  I thought to test the case of no DS3231 installed and found it retuned a valid temperature of 0.2C instead of returning the value of TEMPERATURE_NOT_INSTALLED.  Now it correctly detects the presence of the DS3231 instead of relying on what seemed to be an out-of-range temperature reading which was the original software bug.