Fixed small bug in behaviour of AUTH_LOCK 1
I want to use AUTH_LOCK 1 for NFC tag unlock.  It was impossible for RAPI to unlock the AUTH_LOCK in EVSE_STATE_A.  This was due to an accidental infinite loop of forcing a state change with calling Update(1) instead of Update(0).   Happily this fix does not change the code size by even 1 byte.  Now the behavior described for AUTH_LOCK 1 works as designed.   I fully tested this fix for side effects possibly affecting timers for charging and the fix works perfectly.