Code size optimization: shrink m_GfiRetryCnt and m_NoGndRetryCnt fields
There is a chance this optimization is incorrect, so let me know if so.

Since '255' is the "off" value for these checks, we can assume we'll
never need to track more than 255 retries for either the GFI or No
Ground checks. Thus, use an 8-bit variable for each.

This saves 6 bytes of RAM and 116 bytes of Flash.