Remove need for a second `time` library
Time.cpp and Time.h were used for a single purpose- calling `now()` at
the start and end of the charge and calculating an elapsed charge time
(in seconds).

We already track these same start and end times using our millis()
counter. Calculate the charge time via these instead, and drop the
entire now-unused time library.

The only concern might be "millis increments fast, what about
wraparound?" As long as a charge session isn't taking more than 49.7
days, this should be a non-issue, and I'm not even sure the `now()`
function in the removed time library would have handled it better, since
it used `millis()` under the hood.

https://arduino.stackexchange.com/questions/12587/how-can-i-handle-the-millis-rollover
gives a good description on when you do and don't need to think about
rollover. Since we are doing subtraction, this should not be an issue as
long as the charge session is < 49.7 days.

Before:
    RAM:   [=======   ]  65.4% (used 1339 bytes from 2048 bytes)
    Flash: [========= ]  88.9% (used 29130 bytes from 32768 bytes)

After:
    RAM:   [======    ]  64.6% (used 1323 bytes from 2048 bytes)
    Flash: [========= ]  88.5% (used 29000 bytes from 32768 bytes)

This saves 16 bytes of RAM and 130 bytes of Flash.