Remove g_CurrTime global variable
There are a few global variables that seem to be unnecessary, and this
is one of them. In nearly every case this global was used, it was
initialized with a call to `g_RTC.now()`, and then unpacked into the
component date parts of year/month/day/hour/minute/second.

Replace all usages of this global with a local variable. For now, we
don't touch the global `g_month`, `g_year`, etc. but this might be a
future optimization possiblity.

Before:
    RAM:   [=======   ]  65.7% (used 1345 bytes from 2048 bytes)
    Flash: [========= ]  89.5% (used 29314 bytes from 32768 bytes)

After:
    RAM:   [=======   ]  65.4% (used 1339 bytes from 2048 bytes)
    Flash: [========= ]  88.9% (used 29130 bytes from 32768 bytes)

This change saves 6 bytes of RAM and 184 bytes of Flash.