Fix compilation when SERIALCLI is enabled
I honestly have no idea if anyone is using this anymore. Likely not, as
it hasn't compiled in a bit, but I wanted to do a quick sanity check of
my other changes on this branch. Here's the minimal amount of fixing I
needed to get this to successfully compile.