Recode energy meter usage calculation for better accuracy
For amps and volt measurements with large decimal portions (e.g. 11.9
amps, or 119.9 volts), we were introducing a pretty serious truncation
error in our existing calcUsage function. This is due to using integer
math and the default behavior of truncation, and dividing earlier than
necessary, causing a loss of precision.

The new code is extensively commented to describe what it is doing and
how it works, so hopefully no one is confused in the future with the
rather odd coefficients involved. The important part to note is never
getting too close to the upper limits of what a uint32_t variable can
handle.